+++ Microsoft.Azure.Cosmos/src/Batch/TransactionalBatchOperationResult.cs
@@ -214,7 +214,7 @@ namespace Microsoft.Azure.Cosmos
                 requestMessage: null,
                 headers: headers,
                 cosmosException: null,
-                diagnostics: this.DiagnosticsContext ?? CosmosDiagnosticsContext.Create())
+                diagnostics: this.DiagnosticsContext ?? new CosmosDiagnosticsContextCore())
             {
                 Content = this.ResourceStream
             };