+++ Microsoft.Azure.Cosmos/src/Diagnostics/CosmosDiagnosticsContext.cs
@@ -7,6 +7,8 @@ namespace Microsoft.Azure.Cosmos
     using System.Collections;
     using System.Collections.Generic;
     using Microsoft.Azure.Cosmos.Diagnostics;
+    using Microsoft.Azure.Documents;
+    using static Microsoft.Azure.Cosmos.CosmosClientSideRequestStatistics;
 
     /// <summary>
     /// This represents the diagnostics interface used in the SDK.
@@ -31,6 +33,12 @@ namespace Microsoft.Azure.Cosmos
 
         internal abstract void AddDiagnosticsInternal(QueryPageDiagnostics queryPageDiagnostics);
 
+        internal abstract void AddDiagnosticsInternal(StoreResponseStatistics storeResponseStatistics);
+
+        internal abstract void AddDiagnosticsInternal(AddressResolutionStatistics addressResolutionStatistics);
+
+        internal abstract void AddDiagnosticsInternal(CosmosClientSideRequestStatistics clientSideRequestStatistics);
+
         internal abstract void AddDiagnosticsInternal(CosmosDiagnosticsContext newContext);
 
         internal abstract void SetSdkUserAgent(string userAgent);
@@ -44,12 +52,7 @@ namespace Microsoft.Azure.Cosmos
 
         internal static CosmosDiagnosticsContext Create(RequestOptions requestOptions)
         {
-            return requestOptions?.DiagnosticContext ?? CosmosDiagnosticsContext.Create();
-        }
-
-        internal static CosmosDiagnosticsContext Create()
-        {
-            return new CosmosDiagnosticsContextCore();
+            return requestOptions?.DiagnosticContext ?? new CosmosDiagnosticsContextCore();
         }
     }
 }